@@ -1,15 +1,25 @@
-ActionController::Routing::Routes.draw do |map|
-  map.qunit ':controller/qunit', :action => 'qunit'
+Velocity::Application.routes.draw do
+  match ':controller/qunit', :action => 'qunit', :as => 'qunit'
 
-  map.resources :users do |users|
-    users.resources :tasks, :only => :index
+  resources :users do
+    resources :tasks, :only => :index
   end
-  map.resources :user_sessions
-  map.resources :tasks, :collection => {:verified => :get, :poll => :get}, :member => {:sort => :put}
-  map.resources :releases, :only => [:index, :create]
-  map.resources :stories, :member => {:sort => :put}
-  map.resources :stats, :only => :index, :collection => {:graph_data => :get}
+  resources :user_sessions
+  resources :tasks do
+    collection do
+      get 'verified'
+      get 'poll'
+    end
+    put 'sort', :on => :member
+  end
+  resources :releases, :only => [:index, :create]
+  resources :stories do
+    put 'sort', :on => :member
+  end
+  resources :stats, :only => :index do
+    get 'graph_data', :on => :collection
+  end
 
-  map.redirect '/', :controller => 'tasks'
+  root :to  => 'tasks#index'
 end
 
