@@ -1,4 +1,4 @@
-require 'spec/spec_helper'
+require 'spec_helper'
 
 describe StoriesController do
   before do
@@ -6,14 +6,14 @@ describe StoriesController do
   end
 
   it 'creates a new story' do
-    post :create, :story => {:name => 'Group'}
+    post :create, :story => {:name => 'Group'}, :format => 'js'
     response.should be_success
     Story.count.should == 1
   end
 
   it 'reactivates a previously added story' do
     group = Factory(:story, :name => 'Group', :active => false)
-    post :create, :story => {:name => 'Group'}
+    post :create, :story => {:name => 'Group'}, :format => 'js'
     response.should be_success
     group.reload.active.should be_true
   end
