@@ -1,4 +1,4 @@
-require 'spec/spec_helper'
+require 'spec_helper'
 
 describe Release do
   before do
@@ -55,7 +55,7 @@ describe Release do
     task = Factory(:task, :state => 'verified')
     tasks = []
     5.times {tasks << Factory(:task, :state => 'verified')}
-    
+
     Factory(:release, :created_at => Time.now - 3.days)
     release = Factory(:release, :tasks => [task], :created_at => Time.now - 2.day)
     Factory(:release, :tasks => tasks, :created_at => Time.now)
